fix: modify to build target arm64 and some suggestions in last PR

+++ codec/test/demo/v2.0/include/codec_hdi_encode.h
@@ -88,7 +88,7 @@ public:
     }
     void WaitForStatusChanged();
     void onStatusChanged();
-    bool ReadOneFrame(FILE *fp, char *buf, size_t &filledCount);
+    bool ReadOneFrame(FILE *fp, char *buf, uint32_t &filledCount);
 
 private:
     int32_t OnEmptyBufferDone(const struct OmxCodecBuffer &buffer);
